fix: Open modal only when receiving ack from the websocket
For some obscure reason, when calling `nango.auth()` from a (Rust)
webassembly code, the following error is raised:
```
Uncaught Error: closure invoked recursively or after being dropped
```
on the `window.open()` line 🤷

It used to work on previous Nango implementations when the modal
was open in WebSocket handler.
Reverting to this implementation until the root cause is fixed.